Restructure summary
