+++ app/generator/content/summary/get-business-name.js
@@ -1,0 +1,12 @@
+const getBusinessName = (businessName) => {
+  return {
+    columns: [
+      { width: 200, text: 'Business name:' },
+      { width: '*', text: businessName }
+    ],
+    style: 'column',
+    columnGap: 10
+  }
+}
+
+module.exports = getBusinessName
